user-service/requirements.txt
# requirements.txt
flask
pyjwt

